# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  isAdmin: boolean;
  viewcoins: number;
  createdAt: string;
}

export interface UserPublic {
  id: number;
  username: string;
  viewcoins: number;
}

export interface ViewcoinTransaction {
  id: number;
  userId: number;
  amount: number;
  minutesWatched: number;
  channelName: string;
  earnedAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  viewcoins: number;
  totalMinutesWatched: number;
  createdAt: string;
  recentTransactions: ViewcoinTransaction[];
}

export interface UserUpdate {
  /** @minLength 3 */
  username?: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ScheduleSlot {
  id: number;
  /** 0=Domingo, 1=Segunda, ..., 6=Sábado */
  dayOfWeek: number;
  /** Hora de início (0-23) */
  hourStart: number;
  /** Hora de fim (1-24) */
  hourEnd: number;
  memberName: string;
  channelLink: string;
  createdAt: string;
}

export interface CurrentSlotResponse {
  slot?: ScheduleSlot;
  hasSlot: boolean;
}

export interface SlotInput {
  dayOfWeek: number;
  hourStart: number;
  hourEnd: number;
  memberName: string;
  channelLink: string;
}

export interface EarnInput {
  minutesWatched: number;
  channelName: string;
  /** @nullable */
  slotId?: number | null;
}

export interface EarnResponse {
  viewcoinsEarned: number;
  newBalance: number;
  transaction: ViewcoinTransaction;
}

export interface RankingEntry {
  position: number;
  userId: number;
  username: string;
  viewcoins: number;
  totalMinutesWatched: number;
}