# lib/api-zod/src/generated/types/currentSlotResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleSlot } from './scheduleSlot';

export interface CurrentSlotResponse {
  slot?: ScheduleSlot;
  hasSlot: boolean;
}