# lib/api-zod/src/generated/types/earnResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */
import type { ViewcoinTransaction } from './viewcoinTransaction';

export interface EarnResponse {
  viewcoinsEarned: number;
  newBalance: number;
  transaction: ViewcoinTransaction;
}