# lib/api-zod/src/generated/types/slotInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */

export interface SlotInput {
  dayOfWeek: number;
  hourStart: number;
  hourEnd: number;
  memberName: string;
  channelLink: string;
}