# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  isAdmin: boolean;
  viewcoins: number;
  createdAt: string;
}