# lib/api-zod/src/generated/types/userPublic.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */

export interface UserPublic {
  id: number;
  username: string;
  viewcoins: number;
}