# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 3 */
  username?: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
}