# lib/api-zod/src/generated/types/viewcoinTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ViewCoin App API - Grade de Lives com sistema de Viewcoins
 * OpenAPI spec version: 0.1.0
 */

export interface ViewcoinTransaction {
  id: number;
  userId: number;
  amount: number;
  minutesWatched: number;
  channelName: string;
  earnedAt: string;
}